requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
tortoise-orm==0.20.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
aerich==0.7.2

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0

# Utilities
email-validator==2.1.0.post1
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.27.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
httpx
# Documentation
starlette==0.27.0

# frozenlist==1.3.0
# multidict==6.0.2
chromadb
cuid==0.4
cryptography==43.0.3
google==3.0.0
google-cloud-kms==2.11.0
pymupdf
docx2txt
unstructured
python-magic
python-pptx
# langflow
langchain_community
langchain_google_vertexai
langchain-google-genai
langchain_chroma
